% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{adductCalculator}
\alias{adductCalculator}
\title{Generate neutral mass for different addcuts}
\usage{
adductCalculator(mz = NA, charge = NA, mode = "pos", adduct = NA, primary = T)
}
\arguments{
\item{mz}{A mz value}

\item{charge}{possible charge of the ion}

\item{mode}{ionization mode. Can be pos or neg}

\item{adduct}{possible addcut}

\item{primary}{If TRUE, will use primary adducts otherwise up to 3 chanrges will be used}
}
\description{
Given mz, charge, ionization, adduct information this fucntion will estimate neutral mass for addcuts
}
\details{
if searchMultipleChargeAdducts is set to TRUE, we generate all the selected ions based on adductRules:
primary: \link{M-H}-, \link{M-2H+Na}-, \link{M-2H+K}-, \link{M+Cl}-, \link{M+H}+, \link{M+Na}+, \link{M+K}+, \link{M+NH4}+
extended: We genrated up to three changes state for adduct information for each of them
}
