% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.R
\name{aggregate_results}
\alias{aggregate_results}
\title{Aggregate the partial results of a search engine}
\usage{
aggregate_results(path = "", pattern = "txt", verbose = F)
}
\arguments{
\item{verbose}{Show information about different stages of the processes. Default FALSE}

\item{inputMS2List}{A MS2 list generated by \code{\link{map_features}}}

\item{input_camera}{A CAMERA object (results of preprocessing using CAMERA)}

\item{precursorppm}{Precursor ppm mass deviation for database search}

\item{fragmentppm}{Fragment (MS2) ppm mass deviation for database search}

\item{fragmentabs}{Fragment (MS2) absolute mass deviation for database search}

\item{minPrecursorMass}{Minimum precursor mass to consider. Default 0}

\item{maxPrecursorMass}{Maximum precursor mass to consider NA=all. Default NA}

\item{minPeaks}{The minimum number of MS2 peaks that should be present for an ion to be consider}

\item{maxSpectra}{The maximum number of ions to write (mostly used for testing)}

\item{mode}{Ionization mode. Either pos or neg corresponding to positive and negative, respectively}

\item{adductRules}{What types of adducts to consider. Either primary or extended (see details). Default primary}

\item{outputDir}{Output directory to write individual ions}

\item{searchMultipleChargeAdducts}{Set to TRUE if you want to consider the selected adducts (primary or extended) for unmapped ions or ions without adduct information from CAMERA. Default TRUE}

\item{includeMapped}{Set to TRUE if you want to include mapped ions. Default TRUE}

\item{includeUnmapped}{Set to TRUE if you want to include unmapped ions. Default FALSE}
}
\value{
A zip file containing the parameters needed for all the search engines. Each file inside the zip file includes all the set parameters and MS2 ions and estimated neutral mass and adduct information
}
\description{
Each search engine should ideally return a data.frame. However, it can be the case that the process has been terminated.
If the result folder is set, one can use this function to aggreate the partial results into a larger data.frame.
}
\details{
MetaboRAID attempts to estimate the neutral mass of the molecules to decrease the search space of search engines.
The approach starts with mapping the MS2 parent mz and RT to the CAMERA features. We then search whether that specific feature has been annotated with adducts or isoptopes.
If so, the neutral mass will be estimated based on the adduct information.
However if there is no adduct information in CAMERA results for a particular ion, if searchMultipleChargeAdducts is set to TRUE, we generate all the selected ions based on adductRules:
primary: \link{M-H}-, \link{M-2H+Na}-, \link{M-2H+K}-, \link{M+Cl}-, \link{M+H}+, \link{M+Na}+, \link{M+K}+, \link{M+NH4}+
extended: We genrated up to three changes state for adduct information for each of them
}
\examples{
mapped_ms2s<-map_features(inputMS2s = c("data/Pilot_MS_Control_2_peakpicked.mzML","data/Pilot_MS_Pool_2_peakpicked.mzML"),input_camera = camera_results)
add_c<-map_addcuts(inputMS2List=mapped_ms2s,input_camera=camera_results,
                  precursorppm=10,
                  fragmentppm=20,fragmentabs=0.01,minPrecursorMass=NA,maxPrecursorMass=NA,
                  minPeaks=0,maxSpectra=NA,mode="pos",adductRules="primary",
                  outputDir="general_parameters_3",searchMultipleChargeAdducts=T,
                  includeMapped=T,includeUnmapped=F,verbose=T)


}
