% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_if_conda_is_installed}
\alias{check_if_conda_is_installed}
\title{Checks if a package on conda has been installed}
\usage{
check_if_conda_is_installed(pkg = "metaboraid_package", conda = "auto")
}
\arguments{
\item{pkg}{Name of the package}

\item{conda}{conda binary}
}
\value{
none
}
\description{
Checks if a package on conda has been installed
}
\keyword{internal}
