% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_if_software_exist}
\alias{check_if_software_exist}
\title{Checks if a particular software is installed and throws an exception it does not}
\usage{
check_if_software_exist(
  software_name = NA,
  env = "metaboraid_package",
  conda = "auto"
)
}
\arguments{
\item{software_name}{Name of the software}

\item{env}{conda environment to run the software}

\item{conda}{conda binary}
}
\value{
if failded it will return 1
}
\description{
Checks if a particular software is installed and throws an exception it does not
}
\keyword{internal}
