% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{collect.spectra.lists}
\alias{collect.spectra.lists}
\title{Cluster MS2 parent ions}
\usage{
collect.spectra.lists(spectra, mzabs, mzppm, rtabs)
}
\arguments{
\item{spectra}{A list of MS2s}

\item{mzabs}{Absolute mass deviation between MS2 ions}

\item{mzppm}{ppm differences betweem MS2 ions}

\item{rtabs}{Retention time tolerance}

\item{int.threshold}{The peaks not reaching this intensity will be removed}
}
\value{
A nested list in which the first level indicate the group of MS2s
}
\description{
This function accepts a list of MS2s and cluster the corresponding parent ions ions based on user defined boundaries.
}
\keyword{internal}
