% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tools.R
\name{install_tools}
\alias{install_tools}
\title{Install tools needed by MetaboRAID}
\usage{
install_tools(
  method = c("conda"),
  conda = "auto",
  version = "default",
  cfm_id = FALSE,
  ...
)
}
\arguments{
\item{method}{Which method to use to install the packages (only conda is supported)}

\item{conda}{path to conda binary. Default: auto}

\item{version}{Version of the package to install}

\item{cfm_id}{Set to TRUE if you want to run CFM-ID. Please remember that this is not supported on Windows!}
}
\value{
none
}
\description{
This function allows you to install the tools needed for MetaboRAID.
We use conda to install the tools so make sure you have the latest version conda installed (\url{https://docs.conda.io/projects/conda/en/latest/user-guide/install/index.html}).
}
\details{
At this stage, CSI-FINGERID, Metfrag and Chemdistiller can be installed on Linux, IOS and Windows however, you should install CFM-ID yourself (\url{https://cfmid.wishartlab.com/})

We install everything in metaboraid_package environment. However, CFM-ID will be installed in a different environment.
}
\examples{
install_tools()

}
