% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is_windows}
\alias{is_windows}
\title{Check if Windows is the running OS}
\usage{
is_windows()
}
\value{
a logical value TRUE if Windows is the OS
}
\description{
Check if Windows is the running OS
}
\keyword{internal}
