% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{makeDir}
\alias{makeDir}
\title{Make a directory after performing an existence check}
\usage{
makeDir(dir, overwrite = "never")
}
\arguments{
\item{dir}{Name of directory to create}

\item{overwrite}{A character string: never (default), erase, overwrite}
}
\value{
Path to created directory
}
\description{
Throws an exception if file or directory with same name exist
and overwrite is TRUE.
}
\keyword{internal}
