% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_adducts.R
\name{map_adducts}
\alias{map_adducts}
\title{Annotates MS2 data with adduct information}
\usage{
map_adducts(
  inputMS2List = NA,
  input_camera = NA,
  precursorppm = 10,
  fragmentppm = 20,
  fragmentabs = 0.01,
  minPrecursorMass = NA,
  maxPrecursorMass = NA,
  minPeaks = 0,
  maxSpectra = NA,
  mode = "pos",
  adductRules = "primary",
  outputDir = "general_parameters",
  searchMultipleChargeAdducts = T,
  includeMapped = T,
  includeUnmapped = F,
  verbose = T
)
}
\arguments{
\item{inputMS2List}{A MS2 list generated by \code{\link{map_features}}}

\item{input_camera}{A CAMERA object (results of preprocessing using CAMERA)}

\item{precursorppm}{Precursor ppm mass deviation for database search}

\item{fragmentppm}{Fragment (MS2) ppm mass deviation for database search}

\item{fragmentabs}{Fragment (MS2) absolute mass deviation for database search}

\item{minPrecursorMass}{Minimum precursor mass to consider. Default 0}

\item{maxPrecursorMass}{Maximum precursor mass to consider NA=all. Default NA}

\item{minPeaks}{The minimum number of MS2 peaks that should be present for an ion to be consider}

\item{maxSpectra}{The maximum number of ions to write (mostly used for testing)}

\item{mode}{Ionization mode. Either pos or neg corresponding to positive and negative, respectively}

\item{adductRules}{What types of adducts to consider. Either primary or extended (see details). Default primary}

\item{outputDir}{Output directory to write individual ions}

\item{searchMultipleChargeAdducts}{Set to TRUE if you want to consider the selected adducts (primary or extended) for unmapped ions or ions without adduct information from CAMERA. Default TRUE}

\item{includeMapped}{Set to TRUE if you want to include mapped ions. Default TRUE}

\item{includeUnmapped}{Set to TRUE if you want to include unmapped ions. Default FALSE}

\item{verbose}{Show information about different stages of the processes. Default FALSE}
}
\value{
A zip file containing the parameters needed for all the search engines. Each file inside the zip file includes all the set parameters and MS2 ions and estimated neutral mass and adduct information
}
\description{
This function allows you to annotate MS2 parent ions with addcut information and calculate the neutral mass of the ions.
This function also generates the general parameters for the downstream search engines.
}
\details{
MetaboRAID attempts to estimate the neutral mass of the molecules to decrease the search space of search engines.
The approach starts with mapping the MS2 parent mz and RT to the CAMERA features. We then search whether that specific feature has been annotated with adducts or isoptopes.
If so, the neutral mass will be estimated based on the adduct information.
However if there is no adduct information in CAMERA results for a particular ion, if searchMultipleChargeAdducts is set to TRUE, we generate all the selected ions based on adductRules:
primary: \link{M-H}-, \link{M-2H+Na}-, \link{M-2H+K}-, \link{M+Cl}-, \link{M+H}+, \link{M+Na}+, \link{M+K}+, \link{M+NH4}+
extended: We genrated up to three changes state for adduct information for each of them
}
\examples{

library(CAMERA)
library(metaboraid)
# Read MS1 and MS2 files
ms1_files<-system.file("ms1data",c("X1_Rep1.mzML","X2_Rep1.mzML"),package = "metaboraid")
ms2_files<-system.file("ms2data",c("sample1.mzML","sample2.mzML"),package = "metaboraid")

# mass trace detection
xs <- xcmsSet(ms1_files,method="centWave",ppm=30,peakwidth=c(5,10))

# mass trace matching
xsg <- group(xs)

# convert to CAMERA
xsa <- xsAnnotate(xsg)

# Group mass traces
anF <- groupFWHM(xsa, perfwhm = 0.6)

# Detect isotopes
anI <- findIsotopes(anF, mzabs = 0.01)

# Group using correlation
anIC <- groupCorr(anI, cor_eic_th = 0.75)

# Find adducts
anFA <- findAdducts(anIC, polarity="positive")

# map features and MS2s
mapped_features<-map_features(inputMS2s = ms2_files,input_camera = anFA,ppm = 10,rt = 10)

# Map adducts
mapped_adducts<-map_adducts(inputMS2List=mapped_features,input_camera=anFA,
                            precursorppm=10,
                            fragmentppm=20,fragmentabs=0.01,minPrecursorMass=NA,maxPrecursorMass=NA,
                            minPeaks=10,maxSpectra=10,mode="pos",adductRules="primary",
                            outputDir="general_parameters_4",searchMultipleChargeAdducts=T,
                            includeMapped=T,includeUnmapped=F,verbose=T)
}
