% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_features.R
\name{map_features}
\alias{map_features}
\title{Mapping MS2 to CAMERA}
\usage{
map_features(inputMS2s = NA, input_camera = NA, ppm = 10, rt = 10, verbose = T)
}
\arguments{
\item{inputMS2s}{A character vector containing the absolute paths to the MS2 mzML files.}

\item{ppm}{ppm tolerance between the parent ion and the center of the features. Default 10}

\item{rt}{retention time tolerance between the parent ion and the center of the features in seconds. Default 10}

\item{verbose}{Show information about different stages of the processes. Default FALSE}

\item{camera_results}{A CAMERA object e.g. from running CAMERA on the data}
}
\value{
A list with two named elements (mapped, unmapped). The mapped contain a nested list where each element has a name and contain a list of MS2 spectra for a CAMERA feature with index==name of the list.
The unmapped element contains the crude list of all the unmapped MS2s
}
\description{
This function allows you map MS2 data onto the CAMERA quantification data.
We use this function in downstream to create HyperMS2 objects and map data to addcuts etc.
}
\examples{

library(CAMERA)
library(metaboraid)
# Read MS1 and MS2 files
ms1_files<-system.file("ms1data",c("X1_Rep1.mzML","X2_Rep1.mzML"),package = "metaboraid")
ms2_files<-system.file("ms2data",c("sample1.mzML","sample2.mzML"),package = "metaboraid")

# mass trace detection
xs <- xcmsSet(ms1_files,method="centWave",ppm=30,peakwidth=c(5,10))

# mass trace matching
xsg <- group(xs)

# convert to CAMERA
xsa <- xsAnnotate(xsg)

# Group mass traces
anF <- groupFWHM(xsa, perfwhm = 0.6)

# Detect isotopes
anI <- findIsotopes(anF, mzabs = 0.01)

# Group using correlation
anIC <- groupCorr(anI, cor_eic_th = 0.75)

# Find adducts
anFA <- findAdducts(anIC, polarity="positive")

# map features and MS2s
mapped_features<-map_features(inputMS2s = ms2_files,input_camera = anFA,ppm = 10,rt = 10)

}
