% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{map_id_to_camera}
\alias{map_id_to_camera}
\title{Map identification results onto camera object}
\usage{
map_id_to_camera(
  metFragSearchResult = NA,
  cameraObject = NA,
  ppm = 5,
  MinusTime = 5,
  PlusTime = 5,
  method = "regular",
  ncore
)
}
\arguments{
\item{metFragSearchResult}{A data.frame that much contain the identification results and has to have three addional columns: parentMZ, parentRT and fileName}

\item{cameraObject}{A CAMERA object including addcuts and isotopes}

\item{ppm}{ppm mass devation}

\item{MinusTime}{The lower boundary of the retention time window}

\item{PlusTime}{The higher boundary of the retention time window}

\item{method}{The method for mapping, This can be one of regular, fast, or parallel}

\item{ncore}{Number of cores if parallel is selected for mapping method}
}
\value{
A list of Identification where name of each element correspond to an index of a row in cameraObject
}
\description{
This function accepts a CAMERA object, identification results (a data.frame) and map MSMS data onto the features
}
\keyword{internal}
