% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{merge.spectra.group}
\alias{merge.spectra.group}
\title{Merge MS2 fragment ions}
\usage{
\method{merge}{spectra.group}(spectra, eppm, eabs, int.threshold)
}
\arguments{
\item{spectra}{A list of MS2s}

\item{eppm}{ppm differences betweem MS2 ions}

\item{eabs}{Absolute mass deviation between MS2 ions}

\item{int.threshold}{The peaks not reaching this intensity will be removed}
}
\value{
A list of merged MS2s similar format as input
}
\description{
This function accepts a list of MS2s and cluster the corresponding ions based on user defined boundaries.
}
\keyword{internal}
