% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{parameterToCommand}
\alias{parameterToCommand}
\title{Generates metfrag format parameters}
\usage{
parameterToCommand(param, outputName = "")
}
\arguments{
\item{param}{Parameter list}

\item{outputName}{Name a file for output}
}
\value{
none
}
\description{
Given the list of parameters, this function will generate a file containing those paramters
}
\keyword{internal}
