% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ppmCal}
\alias{ppmCal}
\title{Calculate back the mz difference given ppm}
\usage{
ppmCal(run, ppm)
}
\arguments{
\item{run}{A mz value}

\item{ppm}{ppm difference}
}
\value{
mz difference
}
\description{
Given mz, charge, ionization, adduct information this fucntion will estimate neutral mass for addcuts
}
\keyword{internal}
