% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_msms.R
\name{preprocess_msms}
\alias{preprocess_msms}
\title{Create HyperMS2 from mapped and unmapped MS2s}
\usage{
preprocess_msms(
  mapped_msms = NA,
  centroid = F,
  merge = TRUE,
  centroid_after_merge = F,
  ppm = 10,
  ppm_precursor = 10,
  abs_mz = 0.01,
  abs_mz_precursor = 0.01,
  rt = 10,
  centroid_onlymapped = FALSE,
  merge_onlymapped = FALSE,
  int_threshold = 0,
  verbose = FALSE
)
}
\arguments{
\item{mapped_msms}{A MS2 list generated by \code{\link{map_features}}}

\item{centroid}{If each individual ion should be centroided first. Default TRUE}

\item{merge}{If set to TRUE, MS2s will be merge to HyperMS2. Default TRUE}

\item{centroid_after_merge}{If set to TRUE, a centroiding will be done on each HyperMS2.}

\item{ppm}{ppm tolerance for matching MS2s. Default 10}

\item{ppm_precursor}{ppm tolerance for matching precursors only for unmapped MS2s. Default 10}

\item{abs_mz}{Absolute tolerance for matching MS2s. Default 0.01}

\item{abs_mz_precursor}{Absolute tolerance for matching precursors only for unmapped MS2s. Default 0.01}

\item{rt}{Retention time tolerance for matching precursors only for unmapped MS2s (seconds). Default 10}

\item{centroid_onlymapped}{Set to TRUE to performed centroiding only for mapped MS2s. Default TRUE}

\item{merge_onlymapped}{Set to TRUE to performed merging only for mapped MS2s. Default FALSE}

\item{int_threshold}{The absolute intensity threshold for removing MS2 signals. Default 0}

\item{verbose}{Show information about different stages of the processes. Default FALSE}
}
\value{
A similar format as input but the MS2s has been pre-processed: A list with two elements, mapped and unmapped.
}
\description{
This function is used to merge mapped and unmapped MS2s.
}
\details{
Often multiple ions are picked by mass spectrometer are originated from the same precursor.
This function attempts to combine the ions originating from the same precursor.
For mapped ions, we merge the MS2s that have been mapped to the same feature.
For unmapped ions, we do a clustering using RT and ppm tolerance (on precursor) we then perform the merging.
}
\examples{

library(CAMERA)
library(metaboraid)
# Read MS1 and MS2 files
ms1_files<-system.file("ms1data",c("X1_Rep1.mzML","X2_Rep1.mzML"),package = "metaboraid")
ms2_files<-system.file("ms2data",c("sample1.mzML","sample2.mzML"),package = "metaboraid")

# mass trace detection
xs <- xcmsSet(ms1_files,method="centWave",ppm=30,peakwidth=c(5,10))

# mass trace matching
xsg <- group(xs)

# convert to CAMERA
xsa <- xsAnnotate(xsg)

# Group mass traces
anF <- groupFWHM(xsa, perfwhm = 0.6)

# Detect isotopes
anI <- findIsotopes(anF, mzabs = 0.01)

# Group using correlation
anIC <- groupCorr(anI, cor_eic_th = 0.75)

# Find adducts
anFA <- findAdducts(anIC, polarity="positive")

# map features and MS2s
mapped_features<-map_features(inputMS2s = ms2_files,input_camera = anFA,ppm = 10,rt = 10)
preprocess_msms(mapped_features)


}
