% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{toMetfragCommand}
\alias{toMetfragCommand}
\title{Generate addcut information for MS2 data}
\usage{
toMetfragCommand(
  mappedMS2 = NA,
  unmappedMS2 = NA,
  cameraObject = NA,
  searchMultipleChargeAdducts = F,
  includeUnmapped = T,
  includeMapped = T,
  settingsObject = list(),
  preprocess = NA,
  savePath = "",
  minPeaks = 0,
  maxSpectra = NA,
  maxPrecursorMass = NA,
  minPrecursorMass = NA,
  mode = "pos",
  primary = T
)
}
\arguments{
\item{mappedMS2}{A mapped MS2 list generated by \code{\link{map_features}}}

\item{unmappedMS2}{A unmapped MS2 list generated by \code{\link{map_features}}}

\item{cameraObject}{A CAMERA object (results of preprocessing using CAMERA)}

\item{searchMultipleChargeAdducts}{Set to TRUE if you want to consider the selected adducts (primary or extended) for unmapped ions or ions without adduct information from CAMERA. Default TRUE}

\item{includeUnmapped}{Set to TRUE if you want to include unmapped ions. Default FALSE}

\item{includeMapped}{Set to TRUE if you want to include mapped ions. Default TRUE}

\item{settingsObject}{a list. Can be empty or build using another round of mapping on another file}

\item{minPeaks}{The minimum number of MS2 peaks that should be present for an ion to be consider}

\item{maxSpectra}{The maximum number of ions to write (mostly used for testing)}

\item{maxPrecursorMass}{Maximum precursor mass to consider NA=all. Default NA}

\item{minPrecursorMass}{Minimum precursor mass to consider. Default 0}

\item{mode}{Ionization mode. Either pos or neg corresponding to positive and negative, respectively}

\item{primary}{What types of adducts to consider. If true primary otherwise extended (see details). Default primary}
}
\value{
a list of parameters
}
\description{
Given the list of MS2s (mapped and/or unmapped), this function estimate the neutral mass and attempts to associate addcuts to MS2 ions
}
\details{
MetaboRAID attempts to estimate the neutral mass of the molecules to decrease the search space of search engines.
The approach starts with mapping the MS2 parent mz and RT to the CAMERA features. We then search whether that specific feature has been annotated with adducts or isoptopes.
If so, the neutral mass will be estimated based on the adduct information.
However if there is no adduct information in CAMERA results for a particular ion, if searchMultipleChargeAdducts is set to TRUE, we generate all the selected ions based on adductRules:
primary: \link{M-H}-, \link{M-2H+Na}-, \link{M-2H+K}-, \link{M+Cl}-, \link{M+H}+, \link{M+Na}+, \link{M+K}+, \link{M+NH4}+
extended: We genrated up to three changes state for adduct information for each of them
}
\keyword{internal}
