% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validate.adduct}
\alias{validate.adduct}
\title{Correct strangely looking adducts}
\usage{
validate.adduct(adduct)
}
\arguments{
\item{adduct}{a string containing addcut}
}
\value{
the corrected addcut if needed correction
}
\description{
Correct strangely looking adducts
}
\keyword{internal}
